Add QB64 Super Dark Blue
This theme is based on the color scheme named Super Dark Blue that ships
with QB64.  I took the colors in that theme (which are used to render
code) and mapped them to the bright colors of the terminal, and then I
reduced their brightness to obtain the darker colors.